[VIC] Add tests to Rustbuild
In order to run tests, various parts of rustbuild are cfg'd out,
generally speaking, these are filesystem-related operations and running
commands. Then, rustbuild is run "as normal" and the various steps that
where run are retrieved from the cache and checked against the expected
results.

Note that this means that the current implementation primarily tests
"what" we build, but doesn't actually test that what we build *will*
build. In other words, it doesn't do any form of dependency verification
for any crate. This is possible to implement, but is considered future
work.

This implementation strives to cfg out as little code as possible; it
also does not currently test anywhere near all of rustbuild. The current
tests are also not checked for "correctness," rather, they simply
represent what we do as of this commit, which may be wrong.

A fake fs module is also added to src/bootstrap/lib.rs and used
throughout rustbuild to make it so that most of the code doesn't have to
cfg out create_dir_all and remove_dir_all, etc.

Test cases are drawn from the old implementation of rustbuild, though
the expected results may vary.

+++ src/bootstrap/doc.rs
@@ -17,7 +17,6 @@
 //! Everything here is basically just a shim around calling either `rustbook` or
 //! `rustdoc`.
 
-use std::fs::{self, File};
 use std::io::prelude::*;
 use std::io;
 use std::path::{Path, PathBuf};
@@ -25,6 +24,7 @@ use std::path::{Path, PathBuf};
 use Mode;
 use build_helper::up_to_date;
 
+use fs::{self, File};
 use util::{cp_r, symlink_dir};
 use builder::{Builder, Compiler, RunConfig, ShouldRun, Step};
 use tool::Tool;
@@ -736,6 +736,7 @@ impl Step for UnstableBookGen {
 }
 
 fn symlink_dir_force(src: &Path, dst: &Path) -> io::Result<()> {
+    if cfg!(test) { return Ok(()); }
     if let Ok(m) = fs::symlink_metadata(dst) {
         if m.file_type().is_dir() {
             try!(fs::remove_dir_all(dst));
